requirements.txt
aiocsv==1.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
Bottleneck==1.6.0
build==1.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
docstring_parser==0.17.0
executing==2.2.1
filetype==1.2.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.124.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==2.19.0
google-cloud-vision==3.10.1
google-crc32c==1.7.1
google-genai==1.13.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpc-google-iam-v1==0.14.3
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.9.1
langchain==0.3.25
langchain-classic==1.0.0
langchain-community==0.3.23
langchain-core==0.3.59
langchain-google-genai==2.1.4
langchain-google-vertexai==3.0.1
langchain-text-splitters==0.3.8
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.3.42
loguru==0.7.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numexpr==2.14.1
numpy==2.2.5
orjson==3.10.18
ormsgpack==1.11.0
packaging==24.2
parso==0.8.5
pexpect==4.9.0
pip-tools==7.5.1
platformdirs==4.5.0
pluggy==1.5.0
prompt_toolkit==3.0.52
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
tenacity==9.1.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
validators==0.35.0
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
win32_setctime==1.2.0
xxhash==3.6.0
yarl==1.20.0
zstandard==0.23.0
